backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
openai-whisper
python-multipart
pydantic
python-dotenv
httpx
